refactor: move logger_config.py into utils and update imports
